+++ src/svmClassification.h
@@ -109,7 +109,7 @@ public:
               unsigned int kFoldValue
               );
     
-    int getNumInputs() const override;
+    size_t getNumInputs() const override;
     std::vector<size_t> getWhichInputs() const override;
         
 #ifndef EMSCRIPTEN
